Change parsing_input_index types
This saves 14 bytes of memory

+++ src/types.h
@@ -75,12 +75,12 @@ typedef struct {
  * Structure for transaction information context.
  */
 typedef struct {
-    transaction_t transaction;            /// structured transaction
-    uint8_t signature[MAX_DER_SIG_LEN];   /// transaction input signature encoded in DER
-    uint8_t signing_input_index;          /// The input index currently being signed
-    uint8_t sighash[32];                  /// The sighash being signed
-    size_t parsing_input_index;
-    size_t parsing_output_index;
+    transaction_t transaction;           /// structured transaction
+    uint8_t signature[MAX_DER_SIG_LEN];  /// transaction input signature encoded in DER
+    uint8_t signing_input_index;         /// The input index currently being signed
+    uint8_t sighash[32];                 /// The sighash being signed
+    uint8_t parsing_input_index;
+    uint8_t parsing_output_index;
 } transaction_ctx_t;
 
 /**
